apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
numpy
s2sphere
psycopg[binary]

python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx